refactor: move args to its own file
